testing/web-platform/tests/atomics/window-and-any-worker/wait/good-views/good-views.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
    <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

/*---
description: >
  Test Atomics.wait on arrays that allow atomic operations,
  in an Agent that is allowed to wait.
---*/

function check_done(rs, t) {
  if (rs.length == 7) {
    rs.sort();
    assert_equals(rs[0], "A timed-out");
    assert_equals(rs[1], "B not-equal");
    assert_equals(rs[2], "C not-equal");
    assert_equals(rs[3], "C not-equal");
    assert_equals(rs[4], "C not-equal");
    assert_equals(rs[5], "C not-equal");
    assert_equals(rs[6], "C not-equal");
    t.done();
  }
}

window.addEventListener("load", function() {
  async_test(function(t) {
    let rs = [];
    let w1 = new Worker('wait-on-zero.js');
    let w2 = new Worker('wait-on-fixed.js');
    let w3 = new Worker('wait-on-index.js');

    w1.onmessage = function(e) {
      rs.push("A " + e.data);
      check_done(rs, t);
    };
    w2.onmessage = function(e) {
      rs.push("B " + e.data);
      check_done(rs, t);
    };
    w3.onmessage = function(e) {
      rs.push("C " + e.data);
      check_done(rs, t);
    };

    let sab = new SharedArrayBuffer(1024);
    let view = new Int32Array(sab, 32, 20);

    w1.postMessage([view]);
    w2.postMessage([view]);

    let good_indices = [
        (view) => 0/-1,
        (view) => '-0',
        (view) => view.length - 1,
        (view) => ({ valueOf: () => 0 }),
        (view) => ({ toString: () => '0', valueOf: false }) // non-callable valueOf triggers invocation of toString
    ];

    setTimeout(function() {
        for (let idxGen of good_indices) {
            let idx = idxGen(view);
            view[idx] = 0;
            // Firefox cannot apply clone algorithm on an Object.
            // See https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm
            if (typeof idx == "object") {
                idx = idx.valueOf ? idx.valueOf() : idx.toString();
            }
            Atomics.store(view, idx, 37);
            w3.postMessage([view, idx]);
        }
    }, 500);
  }, "atomics-dedicated-worker-wait-good-views");
  async_test(function(t) {
    let worker = new Worker('good-views-worker.js');
    worker.onmessage = function(e) {
      check_done(e.data[0], t);
    };
    worker.postMessage([]);
  }, "atomics-dedicated-worker-wait-good-views-from-worker");
  async_test(function(t) {
    let worker = new SharedWorker('good-views-shared-worker.js');
    worker.port.start();
    worker.port.onmessage = function(e) {
      check_done(e.data[0], t);
    };
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-wait-good-views-from-shared-worker");
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let rs = [];
      let w1 = new Worker('wait-on-zero.js');
      let w2 = new Worker('wait-on-fixed.js');
      let w3 = new Worker('wait-on-index.js');

      w1.onmessage = function(e) {
        rs.push("A " + e.data);
        check_done(rs, t);
      };
      w2.onmessage = function(e) {
        rs.push("B " + e.data);
        check_done(rs, t);
      };
      w3.onmessage = function(e) {
        rs.push("C " + e.data);
        check_done(rs, t);
      };

      let sab = new SharedArrayBuffer(1024);
      let view = new Int32Array(sab, 32, 20);

      w1.postMessage([view]);
      w2.postMessage([view]);

      let good_indices = [
          (view) => 0/-1,
          (view) => '-0',
          (view) => view.length - 1,
          (view) => ({ valueOf: () => 0 }),
          (view) => ({ toString: () => '0', valueOf: false }) // non-callable valueOf triggers invocation of toString
      ];

      setTimeout(function() {
          for (let idxGen of good_indices) {
              let idx = idxGen(view);
              view[idx] = 0;
              // Firefox cannot apply clone algorithm on an Object.
              // See https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm
              if (typeof idx == "object") {
                  idx = idx.valueOf ? idx.valueOf() : idx.toString();
              }
              Atomics.store(view, idx, 37);
              w3.postMessage([view, idx]);
          }
      }, 500);
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-wait-good-views-from-iframe");
});

</script>
